Fix height being 64 bit instead of 32

+++ os/include/display.h
@@ -6,7 +6,7 @@ typedef struct
 {
     uint32_t* buffer;
     uint32_t width;
-    uint64_t height;
+    uint32_t height;
 } Display;
 
 void initDisplay();
